+++ flutter_secure_storage_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+- Updated code style
+- Add check for secure context, since operations are only allowed with secure context.
+
 ## 2.0.0
 Stable release containing all changes from beta releases, which are:
 - This plugin requires a minimum dart sdk of 3.3.0 or higher and a minimum flutter version of 3.19.0.