+++ flutter_secure_storage_web/CHANGELOG.md
@@ -1,4 +1,4 @@
-## NEXT
+## 2.1.0
 - Updated code style
 - Add check for secure context, since operations are only allowed with secure context.